+++ CHANGELOG.md
@@ -30,6 +30,8 @@
     ([#232](https://github.com/microsoft/ApplicationInsights-Python/pull/232))
 - Add ability to pass custom configuration into instrumentations
     ([#235](https://github.com/microsoft/ApplicationInsights-Python/pull/235))
+- Fix export interval bug
+    ([#237](https://github.com/microsoft/ApplicationInsights-Python/pull/237))
 
 ## [1.0.0b8](https://github.com/microsoft/ApplicationInsights-Python/releases/tag/v1.0.0b8) - 2022-09-26